fix: 성공 응답 코드 열거형 객체를 enums 패키지가 아닌 exception/code 패키지 하위로 이동
